moving matching functions into a seperate module for reusability and using a new syntax for specifying functions in the algo configs
